Cleanup vendor folder after compression
When calling cargo_vendor, we leave a vendor directory inside
of the source folder. During a typical source run the modules
we have are:

    obs_scm
    compress
    cargo_vendor

The issue is that when we run this service *again*, the previous
copy of "vendor" remains in our checked out source, which means
that we end up with vendor inside both the compressed source
tar AND our vendor.tar.

This change cleans up the vendor folder after we complete compression
to avoid this situation, making the output tars cleaner.