boilerplate: finalize logic of how audit could work
How it works:
This binary will try to run `cargo audit --file <pathtolockfile>`.

This binary will read `_service` file and searches for the service
that corresponds to `cargo_vendor` and attempts to audit them.
If `cargo_vendor` service exists, it will use the following params,
if they exist:
1. `cargotoml`
2. `src`, `srctar`, `srcdir`

`cargotoml`:
The `cargotoml` path is used to check if there is also a lockfile to audit
in the path where it resides. Otherwise, it will attempt to regenerate
the lockfile.

`src` `srctar` `srcdir`:
This will be used to check the lockfile of the sources and audit it.

`lockfiles`:
In case additional lockfiles exist to audit from sources.

Vendored tarball:
Since `cargo_vendor` generates a lockfile and also includes it
when generating the vendored tarball, we also check the lockfiles
in those contents as well.

Signed-off-by: Soc Virnyl Estela <uncomfy@uncomfyhalomacro.pl>