fix: decompressed tarball may or may not have a top-level folder
Directory sources by default will also copy the source folder's
top-level folder

This is not true for archived tarballs because they may not have
a top-level folder when the tarball was created.

In this commit, we will use an educated guess that if the extracted
or decompressed tarball is unpacked to the temporary directory,
we will count the number of files and directories. Of course,
if there is ONLY one, it MUST be a folder. If it's a file,
then we will handle and raise an error.

Signed-off-by: Soc Virnyl Estela <uncomfy@uncomfyhalomacro.pl>