fix: sometimes lockfiles can't be found
Lockfiles are weird. The behavior is to generate the lockfile at the parent
dir of the `--manifest-path` parameter of `cargo vendor`, and NOT where
`cargo vendor` was invoked. If it is a subcrate that is part of a workspace,
the lockfile is regenerated from where the workspace manifest is.

For now we do a simple logic to get the lockfile path.  In the future,
a big maybe, we will check if a sub project's manifest file is part of a
workspace so in that way, we can determine where the lockfile is

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>