improvement: better handling on when and when not to vendor
Before, we will have to vendor if and only if the FIRST manifest
contains a dependency or at least a workspace.

This commit changes this by setting `hasdeps` to the value of the
overall extra manifest files if at least one of them contains a
dependency or at least a workspace.

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>