an improvement of commit 24aaa0056
NOTE
Two things to know:
1. If a manifest is not a workspace manifest, it's likely the lockfile
is in the directory of where the manifest is
2. If a manifest is part or a member of a workspace manifest, then it's
likely that the lockfile is on the path of where the workspace manifest
is.

So we just eagerly take all manifest paths from the parameters, and
just check if there are any lockfiles there.
We canonicalize the path first so we won't get any gotchas if we are
looking for the parent path.

And then we only accept if there is `Some` kind of value there and ignore
if there is `None`.

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>