improvements: fix issues with how we respect lockfile
I had this long misconception about `generate-lockfile`.

If we want to respect a lockfile, we have to avoid
generating a new lockfile if it exists before we
pass `--locked` flag to other operations.

If a lockfile does not exist, we call `generate-lockfile`.
Of course, it won't make sense to pass a `--locked` flag.

We will move over to document this in future commits

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>